Completed forms and home page
Completed adding mui to forms and also adding proper react form validation. As well as this we improved the structure in the auth layout, and the app layout. The home page is done now. Of course later we'll work on themes/color control, but right now we're focused on important functionalities with Mui